Modify hub to interpret different message types

+++ message.go
@@ -2,11 +2,11 @@ package main
 
 type Message struct {
 	client      *Client
-	messageType int
+	messageType byte
 	data        []byte
 }
 
-func newMessage(c *Client, t int, d []byte) *Message {
+func newMessage(c *Client, t byte, d []byte) *Message {
 	return &Message{
 		client:      c,
 		messageType: t,
